Revert "More secure default wp-config.php"
Reverting this because we realized it might be a problem with scaled applications, who do not share a common OPENSHIFT_DATA_DIR.
So if an application was scaled, each server would have different tokens and it would definitely wreak havoc.

We've come up with another solution that will be deployed shortly, but wanted to remove this to limit the number of applications using it.

This reverts commit 9ffc8ab362d6f2de3558a0009f22dd69d335a5f1.