Bug 884934
Currently, there is no way for the Origin server to communicate with
(specifically, receive information from) a quickstart application
to display whatever user credentials that the application sets up.

In the long term, we need to create a means to achieve it, but until
then, we write to a well-known location so that console's controllers
can alert the user to this fact.